Implements #174 (#175)
* Implements #174

Thanks for putting this library together.  Here's the implementation I was proposing in #174 to make the function more robust.

* Adds unit tests for snake_to_upper_camel

These tests were added to cover #174 and then some since it appeared
there wasn't complete coverage.

The unit tests were not in the correct place either. So I moved
them to be with all the other parser/base.py tests.